requirements.txt
# SMS Phishing Detection — Requirements
# تثبيت: pip install -r requirements.txt

# Core ML
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Deep Learning (اختياري — احذفه لو مش عايز DL)
tensorflow>=2.13.0

# API Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Telegram Bot
python-telegram-bot>=20.0

# WhatsApp (Twilio) — اختياري
# twilio>=8.0.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0

fastapi
uvicorn
pydantic
scikit-learn
pandas
numpy
streamlit
requests
plotly
python-telegram-bot